Removed old dead code, configured colorful logging for all logging calls, created Modal for tournament code requests. Still WIP
