+++ lblcs_bot/__main__.py
@@ -1,13 +1,19 @@
 import discord
 import os
-from lblcs_bot.bot import Bot
+from bot import Bot
 from dotenv import load_dotenv
+import logging
+
 #Set up environmental variables
 load_dotenv('.env')
 token = os.getenv('TOKEN')
-guild_id = int(os.getenv('GUILD_ID'))
+guild_id = os.getenv('GUILD_ID')
+
 
 #Bot setup
-intents = discord.Intents.all()
-client = Bot(guild_id,intents)
-client.run(token)
+intents = discord.Intents.default()
+client = Bot(guild_id, intents)
+
+#Logging Setup
+
+client.run(token, log_level=logging.INFO, root_logger=True)
